refactor: 사용자 회원가입 책임 이동
