+++ src/main/java/roomescape/auth/service/AuthService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 import roomescape.auth.service.dto.LoginCheckResponse;
 import roomescape.auth.service.dto.LoginRequest;
 import roomescape.auth.service.dto.LoginResponse;
-import roomescape.auth.service.dto.SignUpRequest;
+import roomescape.member.service.dto.SignUpRequest;
 import roomescape.exception.InvalidMemberException;
 import roomescape.exception.UnauthorizedException;
 import roomescape.member.domain.Member;
@@ -33,17 +33,6 @@ public class AuthService {
     private void validatePassword(LoginRequest request, Member member) {
         if (!member.getPassword().equals(request.password())) {
             throw new UnauthorizedException("이메일 또는 비밀번호가 잘못되었습니다.");
-        }
-    }
-
-    public void signUp(SignUpRequest signUpRequest) {
-        validateEmail(signUpRequest.email());
-        memberRepository.save(new Member(signUpRequest.name(), signUpRequest.email(), signUpRequest.password(), Role.GUEST));
-    }
-
-    private void validateEmail(String email) {
-        if (memberRepository.existsByEmail(email)) {
-            throw new InvalidMemberException("이미 가입된 이메일입니다.");
         }
     }
