+++ src/main/java/roomescape/member/controller/MemberController.java
@@ -1,13 +1,17 @@
 package roomescape.member.controller;
 
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
+import jakarta.validation.Valid;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.member.service.dto.SignUpRequest;
 import roomescape.member.service.MemberResponse;
 import roomescape.member.service.MemberService;
 
+import java.net.URI;
 import java.util.List;
 
 @RestController
+@RequestMapping("/members")
 public class MemberController {
     private final MemberService memberService;
 
@@ -15,8 +19,14 @@ public class MemberController {
         this.memberService = memberService;
     }
 
-    @GetMapping("/members")
+    @GetMapping
     public List<MemberResponse> findAllMembers() {
         return memberService.findAll();
+    }
+
+    @PostMapping
+    public ResponseEntity<Void> register(@RequestBody @Valid SignUpRequest signUpRequest) {
+        MemberResponse memberResponse = memberService.create(signUpRequest);
+        return ResponseEntity.created(URI.create("/members/"+memberResponse.id())).build();
     }
 }
