+++ src/main/java/roomescape/member/service/MemberService.java
@@ -1,8 +1,11 @@
 package roomescape.member.service;
 
 import org.springframework.stereotype.Service;
+import roomescape.exception.InvalidMemberException;
 import roomescape.member.domain.Member;
+import roomescape.member.domain.Role;
 import roomescape.member.domain.repository.MemberRepository;
+import roomescape.member.service.dto.SignUpRequest;
 
 import java.util.List;
 
@@ -22,5 +25,17 @@ public class MemberService {
 
     public Member findById(long id) {
         return memberRepository.getById(id);
+    }
+
+    public MemberResponse create(SignUpRequest signUpRequest) {
+        validateEmail(signUpRequest.email());
+        Member member = memberRepository.save(new Member(signUpRequest.name(), signUpRequest.email(), signUpRequest.password(), Role.GUEST));
+        return new MemberResponse(member);
+    }
+
+    private void validateEmail(String email) {
+        if (memberRepository.existsByEmail(email)) {
+            throw new InvalidMemberException("이미 가입된 이메일입니다.");
+        }
     }
 }
