+++ src/test/java/roomescape/auth/controller/AuthControllerTest.java
@@ -9,7 +9,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.context.jdbc.Sql;
 import roomescape.auth.service.dto.LoginRequest;
-import roomescape.auth.service.dto.SignUpRequest;
+import roomescape.member.service.dto.SignUpRequest;
 
 import static org.hamcrest.Matchers.is;
 
@@ -31,7 +31,7 @@ class AuthControllerTest {
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(new SignUpRequest("lini", "lini@email.com", "lini123"))
-                .when().post("/signup")
+                .when().post("/members")
                 .then().log().all();
 
         //when&then
@@ -50,7 +50,7 @@ class AuthControllerTest {
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(new SignUpRequest("lini", "lini@email.com", "lini123"))
-                .when().post("/signup")
+                .when().post("/members")
                 .then().log().all();
 
         //when&then
@@ -61,17 +61,6 @@ class AuthControllerTest {
                 .assertThat().statusCode(200);
     }
 
-    @DisplayName("회원가입 성공 테스트")
-    @Test
-    void signUp() {
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(new SignUpRequest("lini", "lini@email.com", "lini123"))
-                .when().post("/signup")
-                .then().log().all()
-                .assertThat().statusCode(201).header("location", "/login");
-    }
-
     @DisplayName("인증 조회 성공 테스트")
     @Test
     void check() {
@@ -79,7 +68,7 @@ class AuthControllerTest {
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(new SignUpRequest("lini", "lini@email.com", "lini123"))
-                .when().post("/signup")
+                .when().post("/members")
                 .then().log().all();
 
         String token = RestAssured.given().log().all()
